+++ src/arclet/alconna/config.py
@@ -35,6 +35,7 @@ class Namespace:
     )
     to_text: Callable[[Any], str | None] = field(default=lambda x: x if isinstance(x, str) else None)
     converter: Callable[[str | list], DataCollection[Any]] | None = field(default=lambda x: x)
+    compact: bool = field(default=False)
 
     def __eq__(self, other):
         return isinstance(other, Namespace) and other.name == self.name